change return type to size_t

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int number_of_nodes = 0;
+	size_t number_of_nodes = 0;
 
 	while (h)
 	{
